dep_tree: Move dep tree stuff to own file
